--- .idea/libraries/slf4j_jcl_over.xml
@@ -1,11 +1,0 @@
-<component name="libraryTable">
-  <library name="slf4j.jcl.over" type="repository">
-    <properties maven-id="org.slf4j:jcl-over-slf4j:1.7.30" />
-    <CLASSES>
-      <root url="jar://$MAVEN_REPOSITORY$/org/slf4j/jcl-over-slf4j/1.7.30/jcl-over-slf4j-1.7.30.jar!/" />
-      <root url="jar://$MAVEN_REPOSITORY$/org/slf4j/slf4j-api/1.7.30/slf4j-api-1.7.30.jar!/" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES />
-  </library>
-</component>
