+++ .idea/libraries/slf4j_nop.xml
@@ -1,0 +1,11 @@
+<component name="libraryTable">
+  <library name="slf4j.nop" type="repository">
+    <properties maven-id="org.slf4j:slf4j-nop:1.7.30" />
+    <CLASSES>
+      <root url="jar://$MAVEN_REPOSITORY$/org/slf4j/slf4j-nop/1.7.30/slf4j-nop-1.7.30.jar!/" />
+      <root url="jar://$MAVEN_REPOSITORY$/org/slf4j/slf4j-api/1.7.30/slf4j-api-1.7.30.jar!/" />
+    </CLASSES>
+    <JAVADOC />
+    <SOURCES />
+  </library>
+</component>
